+++ app/components/kwiplasz-game/template.hbs
@@ -63,5 +63,5 @@
     {{/if}}
   </div>
 {{/unless}}
-<WebsocketTest @prop={{prop}} @room={{room}}/>
+
