--- app/components/websocket-test/template.hbs
@@ -1,2 +1,0 @@
-<button onClick={{action 'sendButtonPressed'}}>click</button>
-{{prop}}
